MAINT: interpolate/RBF: reduce duplication between _xp and _np backends
