+++ scipy/interpolate/tests/test_rbfinterp.py
@@ -7,7 +7,7 @@ from scipy.stats.qmc import Halton
 from scipy.spatial import cKDTree  # type: ignore[attr-defined]
 from scipy.interpolate._rbfinterp import (
     _AVAILABLE, _SCALE_INVARIANT, _NAME_TO_MIN_DEGREE, RBFInterpolator,
-    _get_backend
+    _get_backend, _monomial_powers,
     )
 from scipy.interpolate import _rbfinterp_pythran
 from scipy._lib._testutils import _run_concurrent_barrier
@@ -19,7 +19,7 @@ def _vandermonde(x, degree, xp=np):
     # Returns a matrix of monomials that span polynomials with the specified
     # degree evaluated at x.
     backend = _get_backend(xp)
-    powers = backend._monomial_powers(x.shape[1], degree, xp)
+    powers = _monomial_powers(x.shape[1], degree, xp)
     return backend.polynomial_matrix(x, powers, xp)
